requirements.txt
PySide6>=6.6
pyqtgraph>=0.13
PyYAML>=6.0
python-snap7>=1.3
